Set html_side_margin to 0.5 #183449

+++ action.py
@@ -67,7 +67,7 @@ class ActionReport(metaclass=PoolMeta):
         },
         help='If set, a zip file with a document per record will be created.')
     html_copies = fields.Integer('Number of copies')
-    html_side_margin = fields.Integer('Side Margin (cm)')
+    html_side_margin = fields.Float('Side Margin (cm)')
     html_extra_vertical_margin = fields.Integer('Extra Vertical Margin (cm)')
     html_file_name = fields.Char('Report File Name Pattern', translate=True,
         help='By default, the file name is generated from the report name and '
